MAINT: Highlight の更新などをフックに分離
